requirements.txt
fastapi
uvicorn
tensorflow
pandas
numpy
scikit-learn
sentence-transformers
python-dotenv

